Support binary SCCS (^Af e 1)
As requested and indicated by Maxime Fiandino

http://www.gnu.org/software/cssc/manual/Global-Flags-Section.html#Global-Flags-Section

The ‘e’ flag, if set to a nonzero value, indicates that the controlled
file is binary and is therefore stored in uuencoded form in the file body.
If this flag is set to zero or is missing, then the file body is not
encoded. See Flags for information about the other possible flag letters
and their meanings. See Interoperability for information about sharing
sccs files with other implementations of sccs.

The ‘e’ flag is a boolean flag but is stored within the sccs file with a
value, as shown in the example above. When cssc initially writes the sccs
file header for a new sccs fiel created with admin -i, it does not know if
the initial body of the file is binary or not, so ‘^Af f e 0’ is written
into the header and if the file turns out to need encoding, admin will
seek back to the header and change ‘^Af f e 0’ to ‘^Af f e 1’. If binary
file support is disabled (see Binary File Support, ‘^Af f e 0’ is still
used but will never be changed to ‘^Af f e 1’.